pyproject.toml
[project]
name = "booking-monitor"
version = "0.1.0"
description = "Robust iCal booking monitor scaffold with Discord notifications."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "httpx>=0.27,<1.0",
  "icalendar>=6.1,<7.0",
  "PyYAML>=6.0,<7.0",
]

[project.scripts]
booking-monitor = "booking_monitor.app:main"

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/booking_monitor"]

[dependency-groups]
dev = [
  "pytest>=8.3,<9.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
